+++ app/controllers/active_waiter/application_controller.rb
@@ -1,4 +1,11 @@
 module ActiveWaiter
   class ApplicationController < ActionController::Base
+    layout :active_waiter_layout
+
+    private
+
+      def active_waiter_layout
+        ActiveWaiter.configuration.layout
+      end
   end
 end