+++ test/active_waiter/configuration_test.rb
@@ -1,0 +1,19 @@
+require 'test_helper'
+
+class ConfigurationTest < Minitest::Test
+  def teardown
+    ActiveWaiter.configuration = nil
+  end
+
+  def test_configuration_defaults
+    assert_equal "active_waiter/layouts/application", ActiveWaiter.configuration.layout
+  end
+
+  def test_configuration_layout
+    ActiveWaiter.configure do |config|
+      config.layout = 'layouts/application'
+    end
+
+    assert_equal "layouts/application", ActiveWaiter.configuration.layout
+  end
+end