+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Unreleased
 
+- Add configuration option for layout [PR#1](https://github.com/choonkeat/active_waiter/pull/1/)
+
 ## 0.1.0 - 2015.05.24
 
-- Change to toggle download (or keep redirect) in controller/view [0dc1f1](https://github.com/choonkeat/active_waiter/commit/0dc1f1eea6ec6bb9fc5632ce976855d668ad423a)
+- Add toggle for download or redirect in controller/view [0dc1f1](https://github.com/choonkeat/active_waiter/commit/0dc1f1eea6ec6bb9fc5632ce976855d668ad423a)
 
 ## 0.0.1 - 2015.05.24